requirements.txt
# Django Core
Django==5.2.6
django-environ==0.11.2
whitenoise==6.8.2
djangorestframework==3.16.1
django-compressor==4.5.1

# API Documentation
drf-yasg==1.21.7

# WebSocket & Real-time
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Security & Production
python-decouple==3.8
dj-database-url==2.1.0
django-redis==5.4.0
gunicorn==21.2.0

# Storage
boto3==1.29.7
django-storages==1.14.2

# Cache
django-redis==5.4.0

# Push Notifications
pywebpush==1.14.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0
psutil==5.9.6

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Dependencies
requests==2.31.0
pytz==2023.3
pillow==10.1.0

# Analysis & ML
textblob==0.17.1
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.4
numpy>=1.24.3,<1.25.0
joblib==1.3.2

# Development Tools (comentar em produção)
django-extensions==3.2.3

# Advanced Features
pandas==2.1.4
numpy==1.24.4

# Push Notifications
pywebpush==1.14.0
